Updates to database, models and views

+++ app/views.py
@@ -1,6 +1,7 @@
 from flask import Flask, request, jsonify, abort, make_response
-from .models import Users, Questions, Answers
+from app.models import Users, Questions, Answers
 from passlib.hash import sha256_crypt
+
 
 app = Flask(__name__)
 
@@ -25,7 +26,13 @@ def create_a_user_accout():
     email = str(data.get("email"))
     password = sha256_crypt.encrypt(str(data.get("password")))
     if request.method == 'POST':
-        return jsonify(user.add_user_account(user_id, firstname, surname, email, password)), 201
+        try:
+            if not data or data == "None" or len(data) == 0:
+                return make_response(jsonify("REQUIRED FIELD: Don't leave blank or submit spaces!")), 400
+            else:
+                return jsonify(user.add_user_account(user_id, firstname, surname, email, password)), 201
+        except:
+            return make_response(jsonify("User_ID/email already exists! Try again")), 400
 
 @app.route('/login', methods=['POST'])
 def login_a_user():
@@ -33,15 +40,21 @@ def login_a_user():
     user_id = data.get("user_id")
     password = data.get("password")
     if request.method == 'POST':
-        return jsonify(user.login_user_account(user_id, password)), 200
+        if user_id in user.users.keys():
+            if sha256_crypt.verify(password, user.users[user_id]["password"]):
+                return make_response(jsonify("User: %s has logged in successfully!" % (user_id))), 200
+            else:
+                return make_response(jsonify("Password is incorrect, try again")), 400
+        else:
+            return make_response(jsonify("Invalid username: Username doesn't exist!")), 400
 
 @app.route('/questions', methods=['GET'])
 def view_all_questions():
     if request.method == 'GET':
-        """if len(stack.view_questions().keys()) <= 0:
+        if len(stack.view_questions().keys()) <= 0:
             return make_response(jsonify("No available questions to display")), 404
-        else:"""
-        return jsonify(stack.view_questions()), 200
+        else:
+            return jsonify(stack.view_questions()), 200
 
 @app.route('/questions', methods=['POST'])
 def add_question():
@@ -49,39 +62,49 @@ def add_question():
     user_id = str(data.get("user_id"))
     title = str(data.get("title"))
     description = str(data.get("description"))
+    error = "Unable to add question due to missing/duplicate required fields. Try again."
     if request.method == 'POST':
-        """if question.isspace() or question == "None" or len(question) <=0:
-            return make_response(jsonify("REQUIRED FIELD: Don't leave blank or submit spaces!")), 400
-        elif question.isnumeric() :
-            return make_response(jsonify("Invalid Input, please try again!")), 400
-        else:"""
-        return jsonify(stack.add_questions(user_id, title, description)), 201
+        try:
+            if not data or data == "None" or len(data) == 0:
+                return make_response(jsonify("REQUIRED FIELD: Don't leave blank or submit spaces!")), 400
+            else:
+                return jsonify(stack.add_questions(user_id, title, description)), 201
+        except:
+            return make_response(jsonify(error)), 400
 
 @app.route('/questions/<int:questionid>', methods=['GET'])
 def view_a_question(questionid):
     if request.method == 'GET':
-        return jsonify(stack.view_question(questionid)), 200
+        try:
+            return jsonify(stack.view_question(questionid)), 200
+        except:
+            return make_response(jsonify("Question doesn't exist: Check questionID!")), 400
 
 @app.route('/questions/<int:questionid>', methods=['DELETE'])
 def delete_a_question(questionid):
     if request.method == 'DELETE':
-        return jsonify(stack.delete_question(questionid)), 202
+        if questionid in stack.questions.keys():
+            return jsonify(stack.delete_question(questionid)), 202
+        else:
+            return make_response(jsonify("Unable to delete question which doesn't exist: Check questionID!")), 400
         
 @app.route('/questions/<int:questionid>/answers', methods=['POST'])
 def add_an_answer(questionid):
     data = request.get_json()
+    error = "Unable to add answer due to missing/duplicate required fields. Try again."
     if request.method == 'POST':
-        user_id = str(data.get("user_id"))
-        title = str(data.get("title"))
-        description = str(data.get("description"))
-        """if questionid not in stack.questions.keys():
-            return make_response(jsonify("Question not found: Question ID out of range!")), 404
-        elif answer.isdigit():
-            return make_response(jsonify("Invalid Input, please try again!")), 400
-        elif (answer == None) or (len(answer) <= 0) or answer.isspace():
-            return make_response(jsonify("REQUIRED FIELD: Don't leave blank or submit spaces!")), 400
-        else:"""
-        return jsonify(ans.add_answer(questionid, user_id, title, description)), 201
+        try:
+            user_id = str(data.get("user_id"))
+            title = str(data.get("title"))
+            description = str(data.get("description"))
+            if questionid not in stack.questions.keys():
+                return make_response(jsonify("Question not found: Question ID out of range!")), 404
+            elif (data == None) or (len(data) <= 0) or not data:
+                return make_response(jsonify("REQUIRED FIELD: Don't leave blank or submit spaces!")), 400
+            else:
+                return jsonify(ans.add_answer(questionid, user_id, title, description)), 201
+        except:
+            return make_response(jsonify(error)), 400
     else:
         abort(405)
 
@@ -89,15 +112,16 @@ def add_an_answer(questionid):
 def content_not_found(e):
     return make_response(jsonify("Failed to locate questionID with requested answer")), 404
 
-@app.route('/questions/<int:questionid>/answers', methods=['GET'])
-def view_answers(questionid):
-    if request.method == 'GET':
-        return jsonify(ans.view_answers(questionid)), 200
-
 @app.route('/questions/<int:questionid>/answers/<int:answerid>', methods=['POST'])
 def set_as_preferred_answer(questionid, answerid):
     if request.method == 'POST':
-        return jsonify(ans.select_preferred_answer(answerid)), 201
+        if questionid in stack.questions.keys():
+            if answerid in ans.answers.keys():
+                return jsonify(ans.select_preferred_answer(answerid)), 201
+            else:
+                return make_response(jsonify("Unable to locate answer: Check answerID")), 400
+        else:
+            return make_response(jsonify("Failed to locate questionID for requested answer")), 400
 
 if __name__ == '__main__':
     app.run(debug=True)
